DEJ Status Monitor/1.0.0/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy>=2.0.36
apscheduler==3.10.4
httpx==0.25.2
jinja2==3.1.2
python-multipart==0.0.6
python-dotenv==1.0.0
pytest==7.4.3
pytz==2023.3
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4